Add files via upload
Here is a pdf / ppt of an Assessment I did for a company. Here I needed to,
-	Get a quick view of share of total spends and F&B spends for existing customers (restaurants) present in the Database
-	What are the adjacent opportunities from the non-commercial establishments to additionally sell into.
-	Also getting a view of Retailers presence adjacent to the customers and prospects that might help us understand how much of these opportunities are being met / unmet.

I had to prepare a ppt where in I was supposed to highlight these points and had to provide suggestions / recommendations for the same. I had used Microsoft Power BI inorder to analyze the data and to create visuals which are present in the ppt/pdf.